Implement contribution feedback: use external hook to save addresses and bugfix
