+++ src/utils/index.js
@@ -1,8 +1,4 @@
-import RootElement from '../../../../utils/rootElement';
-
 export const parseAddress = (amazonAddress, cartId) => {
-  const addressFields = RootElement.getAddressConfig();
-  const prefixSelected = addressFields?.prefix?.options[0].value ?? 'Female';
   const street =
     amazonAddress.street.length > 1
       ? [amazonAddress.street.join(', ')]
@@ -20,7 +16,6 @@ export const parseAddress = (amazonAddress, cartId) => {
     region: amazonAddress.region_code,
     street,
     zipcode: amazonAddress.postcode,
-    prefix: prefixSelected,
     isSameAsShipping: amazonAddress.isSameAsShipping,
   };
 };
